Various line/import cleanup and move AddressState to core-package
