+++ src/main/java/org/ethereum/gui/WalletAddressPanel.java
@@ -1,13 +1,14 @@
 package org.ethereum.gui;
 
+import org.ethereum.core.AddressState;
 import org.ethereum.util.Utils;
-import org.ethereum.wallet.AddressState;
 import org.spongycastle.util.encoders.Hex;
 
 import javax.swing.*;
 import javax.swing.border.Border;
 import javax.swing.border.CompoundBorder;
 import javax.swing.border.EmptyBorder;
+
 import java.awt.*;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
